app/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.28.0
pydantic==2.6.4
sqlalchemy==2.0.28
asyncpg==0.29.0
python-multipart==0.0.9
python-dotenv==1.0.1
qdrant-client==1.8.1
motor==3.3.2
pymongo==4.6.2
pytz==2024.1
openai==1.14.3
